wizling/requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.1.2
django-classy-tags==4.1.0
django-cms==4.1.3
django-entangled==0.5.4
django-filer==3.2.3
django-filter==24.3
django-formtools==2.5.1
django-fsm==2.8.2
django-modelcluster==6.3
django-parler==2.3
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-taggit==6.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-form-builder==0.1.1
djangocms-frontend==1.3.4
djangocms-icon==2.1.0
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.6
djangocms-versioning==2.1.0
djangorestframework==3.15.2
draftjs-exporter==5.0.0
easy-thumbnails==2.10
et_xmlfile==2.0.0
fastapi==0.115.4
filetype==1.2.0
Flask==3.0.3
graphql-core==3.2.5
h11==0.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
l18n==2021.3
laces==0.1.1
lxml==5.3.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
openpyxl==3.1.5
packaging==24.1
pillow==11.0.0
pillow_heif==0.20.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gql==0.2.4
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.2
starlette-graphql==0.2.1
strawberry-graphql==0.247.0
svglib==1.5.1
telepath==0.3.1
tinycss2==1.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wagtail==6.3
webencodings==0.5.1
Werkzeug==3.1.0
Willow==1.9.0
django-cors-headers==4.3.1 

